requirements.txt
openai
anthropic
requests
google-generativeai
python-dotenv
python-socketio[client]
websocket-client
python-socketio
